refactor(admin): extract useTheme/useLocale/useFocusTrap hooks from AdminLayout
Addresses PR #119 review comment r3107533744 from @Jorenve.

- useTheme: encapsulates resolveTheme, applyTheme, theme state
- useLocale: encapsulates getInitialLocale, handleLocaleChange, locale state
- useFocusTrap: encapsulates focus-trap keydown effect + initial-focus effect;
  accepts containerRef and active flag
- logoutAndRedirect: moved to src/auth/adminLogout.ts (co-located with auth
  utilities) — pure async function, no React state
